Refactor edit component into smaller pieces
