+++ src/blocks/bar/index.js
@@ -11,10 +11,8 @@ const { registerBlockType } = wp.blocks;
 /**
  * Components and dependencies.
  */
-import { ChartStyles, DataStyles } from './components';
-import { Bar } from 'react-chartjs-2';
-import { Edit, Save } from '../../common/components';
-import { randomColors, randomValues } from '../../common/helpers';
+import { Edit } from './components';
+import { Save } from '../../common/components';
 
 /**
  * Registers this as a block.
@@ -130,38 +128,7 @@ registerBlockType( 'hello-charts/block-bar', {
 
 	/* Render the block components. */
 	edit: ( props ) => {
-		const {
-			attributes: {
-				blockId,
-				chartData,
-				chartOptions,
-				height,
-				width,
-			},
-		} = props;
-
-		const parsedData = JSON.parse( chartData );
-		const parsedOptions = JSON.parse( chartOptions );
-
-		return (
-			<Edit
-				{ ...props }
-				ChartStyles={ ChartStyles }
-				DataStyles={ DataStyles }
-				chartType="bar"
-				maybeGenerateData={ maybeGenerateData }
-				onNewDataset={ onNewDataset }
-				titlePlaceholder={ __( 'Bar Chart', 'hello-charts' ) }
-			>
-				<Bar
-					height={ height }
-					width={ width }
-					id={ blockId }
-					data={ parsedData }
-					options={ parsedOptions }
-				/>
-			</Edit>
-		);
+		return <Edit { ...props } />;
 	},
 
 	/* Save the block markup. */
@@ -169,25 +136,3 @@ registerBlockType( 'hello-charts/block-bar', {
 		return <Save { ...props } />;
 	},
 } );
-
-const maybeGenerateData = ( datasets ) => {
-	const themeColors = randomColors( datasets.length );
-
-	datasets.forEach( ( dataset, index ) => {
-		if ( 'generate' === dataset.data[ 0 ] ) {
-			dataset.data = randomValues( 8 );
-		}
-
-		if ( ! dataset.hasOwnProperty( 'backgroundColor' ) ) {
-			dataset.backgroundColor = themeColors[ index ];
-		}
-	} );
-};
-
-const onNewDataset = ( dataset ) => {
-	const color = randomColors( 1 ).shift();
-
-	dataset.label = __( 'New Data Set', 'hello-charts' );
-	dataset.backgroundColor = color;
-};
-
