+++ src/blocks/line/index.js
@@ -11,10 +11,8 @@ const { registerBlockType } = wp.blocks;
 /**
  * Components and dependencies.
  */
-import { ChartStyles, DataStyles } from './components';
-import { Line } from 'react-chartjs-2';
-import { Edit, Save } from '../../common/components';
-import { hex2rgba, randomColors, randomValues } from '../../common/helpers';
+import { Edit } from './components';
+import { Save } from '../../common/components';
 
 /**
  * Registers this as a block.
@@ -170,38 +168,7 @@ registerBlockType( 'hello-charts/block-line', {
 
 	/* Render the block components. */
 	edit: ( props ) => {
-		const {
-			attributes: {
-				blockId,
-				chartData,
-				chartOptions,
-				height,
-				width,
-			},
-		} = props;
-
-		const parsedData = JSON.parse( chartData );
-		const parsedOptions = JSON.parse( chartOptions );
-
-		return (
-			<Edit
-				{ ...props }
-				ChartStyles={ ChartStyles }
-				DataStyles={ DataStyles }
-				chartType="line"
-				maybeGenerateData={ maybeGenerateData }
-				onNewDataset={ onNewDataset }
-				titlePlaceholder={ __( 'Line Chart', 'hello-charts' ) }
-			>
-				<Line
-					height={ height }
-					width={ width }
-					id={ blockId }
-					data={ parsedData }
-					options={ parsedOptions }
-				/>
-			</Edit>
-		);
+		return <Edit { ...props } />;
 	},
 
 	/* Save the block markup. */
@@ -209,28 +176,3 @@ registerBlockType( 'hello-charts/block-line', {
 		return <Save { ...props } />;
 	},
 } );
-
-const maybeGenerateData = ( datasets ) => {
-	const themeColors = randomColors( datasets.length );
-
-	datasets.forEach( ( dataset, index ) => {
-		if ( 'generate' === dataset.data[ 0 ] ) {
-			dataset.data = randomValues( 6 );
-		}
-
-		if ( ! dataset.hasOwnProperty( 'backgroundColor' ) ) {
-			dataset.borderColor = themeColors[ index ];
-			dataset.pointBackgroundColor = themeColors[ index ];
-			dataset.backgroundColor = hex2rgba( themeColors[ index ], 0.6 );
-		}
-	} );
-};
-
-const onNewDataset = ( dataset ) => {
-	const color = randomColors( 1 ).shift();
-
-	dataset.label = __( 'New Data Set', 'hello-charts' );
-	dataset.borderColor = color;
-	dataset.pointBackgroundColor = color;
-	dataset.backgroundColor = hex2rgba( color, 0.6 );
-};
