+++ src/blocks/pie/components/index.js
@@ -1,2 +1,3 @@
 export { default as ChartStyles } from './chart-styles';
 export { default as DataStyles } from './data-styles';
+export { default as Edit } from './edit';
