+++ src/blocks/radar/index.js
@@ -11,10 +11,8 @@ const { registerBlockType } = wp.blocks;
 /**
  * Components and dependencies.
  */
-import { ChartStyles, DataStyles } from './components';
-import { Radar } from 'react-chartjs-2';
-import { Edit, Save } from '../../common/components';
-import { hex2rgba, randomColors, randomValues } from '../../common/helpers';
+import { Edit } from './components';
+import { Save } from '../../common/components';
 
 /**
  * Registers this as a block.
@@ -159,38 +157,7 @@ registerBlockType( 'hello-charts/block-radar', {
 
 	/* Render the block components. */
 	edit: ( props ) => {
-		const {
-			attributes: {
-				blockId,
-				chartData,
-				chartOptions,
-				height,
-				width,
-			},
-		} = props;
-
-		const parsedData = JSON.parse( chartData );
-		const parsedOptions = JSON.parse( chartOptions );
-
-		return (
-			<Edit
-				{ ...props }
-				ChartStyles={ ChartStyles }
-				DataStyles={ DataStyles }
-				chartType="radar"
-				maybeGenerateData={ maybeGenerateData }
-				onNewDataset={ onNewDataset }
-				titlePlaceholder={ __( 'Radar Chart', 'hello-charts' ) }
-			>
-				<Radar
-					height={ height }
-					width={ width }
-					id={ blockId }
-					data={ parsedData }
-					options={ parsedOptions }
-				/>
-			</Edit>
-		);
+		return <Edit { ...props } />;
 	},
 
 	/* Save the block markup. */
@@ -198,29 +165,3 @@ registerBlockType( 'hello-charts/block-radar', {
 		return <Save { ...props } />;
 	},
 } );
-
-const maybeGenerateData = ( datasets ) => {
-	const themeColors = randomColors( datasets.length );
-
-	datasets.forEach( ( dataset, index ) => {
-		if ( 'generate' === dataset.data[ 0 ] ) {
-			dataset.data = randomValues( 7 );
-		}
-
-		if ( ! dataset.hasOwnProperty( 'backgroundColor' ) ) {
-			dataset.borderColor = themeColors[ index ];
-			dataset.pointBackgroundColor = themeColors[ index ];
-			dataset.backgroundColor = hex2rgba( themeColors[ index ], 0.6 );
-		}
-	} );
-};
-
-const onNewDataset = ( dataset ) => {
-	const color = randomColors( 1 ).shift();
-
-	dataset.label = __( 'New Data Set' );
-	dataset.borderColor = color;
-	dataset.pointBackgroundColor = color;
-	dataset.backgroundColor = hex2rgba( color, 0.6 );
-};
-
