+++ src/common/components/index.js
@@ -1,5 +1,5 @@
+export { default as ChartBlock } from './chart-block';
 export { default as ChartFormattingToolbar } from './chart-formatting-toolbar';
-export { default as Edit } from './edit';
 export { default as EditDataButton } from './edit-data-button';
 export { default as EditDataModal } from './edit-data-modal';
 export { default as EditDataToolbar } from './edit-data-toolbar';
